Add new FT for sharing
Also move session creation stuff to base
