+++ src/functional_tests/base.py
@@ -1,13 +1,15 @@
 import os
 import time
 
+from django.conf import settings
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 from selenium import webdriver
 from selenium.common.exceptions import WebDriverException
 from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
 
-from .container_commands import reset_database
+from .container_commands import create_session_on_server, reset_database
+from .management.commands.create_session import create_pre_authenticated_session
 
 
 MAX_WAIT = 10
@@ -41,6 +43,22 @@ class FunctionalTest(StaticLiveServerTestCase):
 
     def tearDown(self):
         self.browser.quit()
+
+    def create_pre_authenticated_session(self, email):
+        if self.test_server:
+            session_key = create_session_on_server(self.test_server, email)
+        else:
+            session_key = create_pre_authenticated_session(email)
+        ## to set a cookie we need to first visit the domain.
+        ## 404 pages load the quickest!
+        self.browser.get(self.live_server_url + "/404_no_such_url/")
+        self.browser.add_cookie(
+            dict(
+                name=settings.SESSION_COOKIE_NAME,
+                value=session_key,
+                path="/",
+            )
+        )
 
     @wait
     def wait_for(self, fn):
